+++ tests/testthat/test-block.R
@@ -3,7 +3,7 @@ context("test-block.R")
 test_that("block yaml works", {
   expect_error(block_yaml(license = "foo"), msg = "license not legal")
   expect_identical(
-    block_config(),
+    vw_block_config(),
     "license: mit\nheight: 500\nscrolling: yes\nborder: yes\n"
   )
 })